Refactor getTaskInput
